♻️ refactor: Alterando detalhes dos códigos

+++ 1-EstruturaSequencial/ex05/src/ex05.java
@@ -4,7 +4,7 @@ public class ex05{
    public static void main(String[] args) {
       try(Scanner t = new Scanner(System.in)){
          System.out.print("digite um valor em metros: ");
-         float metros = t.nextFloat();
+         double metros = t.nextFloat();
          double centimetros = metros*100;
          System.out.print(metros + "m = " + centimetros + "cm");
       }
